userEntry removal, CSV output
In order for quiz submissions (cuePoints of a type quiz.QUIZ_ANSWER) to be removed, the record of their submission has to  be eliminated as well. 

The script now includes a userEntry.delete action on all objects related to user submissions. 

The script also now exports a .csv containing detailed information about what was deleted. The document is time stamped.